remove @celo/dev-util package It will be moved to the developer tooling repo as its more used there than here.
The one function that is used here has been lined in the place its used.
