--- packages/dev-utils/package.json
@@ -1,40 +1,0 @@
-{
-  "name": "@celo/dev-utils",
-  "version": "0.0.2",
-  "private": false,
-  "description": "util package for celo packages that should only be a devDependency",
-  "main": "./lib/index.js",
-  "types": "./lib/index.d.ts",
-  "author": "Celo",
-  "license": "Apache-2.0",
-  "homepage": "https://github.com/celo-org/celo-monorepo/tree/master/packages/dev-utils",
-  "repository": "https://github.com/celo-org/celo-monorepo/tree/master/packages/dev-utils",
-  "keywords": [
-    "celo"
-  ],
-  "files": [
-    "lib/*"
-  ],
-  "scripts": {
-    "build": "tsc -b .",
-    "lint": "yarn run --top-level eslint .",
-    "prepack": "yarn build"
-  },
-  "dependencies": {
-    "ganache": "npm:@celo/ganache@7.8.0-unofficial.0",
-    "bignumber.js": "^9.0.0",
-    "web3": "1.10.0",
-    "web3-core-helpers": "1.10.0",
-    "tmp": "^0.1.0",
-    "targz": "^1.0.1",
-    "fs-extra": "^8.1.0"
-  },
-  "devDependencies": {
-    "@tsconfig/recommended": "^1.0.3",
-    "@types/fs-extra": "^8.1.0",
-    "@types/targz": "1.0.4"
-  },
-  "engines": {
-    "node": ">=18.14.2"
-  }
-}
