--- packages/dev-utils/src/describeEach.ts
@@ -1,9 +1,0 @@
-export interface TestCase {
-  label: string
-}
-
-export function describeEach<T extends TestCase>(testCases: T[], fn: (testCase: T) => void) {
-  for (const testCase of testCases) {
-    describe(testCase.label, () => fn(testCase))
-  }
-}
