--- packages/dev-utils/src/migration-override.json
@@ -1,127 +1,0 @@
-{
-  "downtimeSlasher": {
-    "slashableDowntime": 60
-  },
-  "epochRewards": {
-    "frozen": false
-  },
-  "exchange": {
-    "frozen": false,
-    "minimumReports": 1
-  },
-  "exchangeEUR": {
-    "frozen": false,
-    "minimumReports": 1
-  },
-  "goldToken": {
-    "frozen": false
-  },
-  "governance": {
-    "dequeueFrequency": 30,
-    "queueExpiry": 1000,
-    "approvalStageDuration": 100,
-    "referendumStageDuration": 100,
-    "executionStageDuration": 100,
-    "minDeposit": 1,
-    "concurrentProposals": 5,
-    "skipTransferOwnership": false
-  },
-  "governanceApproverMultiSig": {
-    "signatories": [
-      "0x5409ed021d9299bf6814279a6a1411a7e866a631"
-    ],
-    "numRequiredConfirmations": 1
-  },
-  "grandaMento": {
-    "approver": "0x5409ED021D9299bf6814279A6A1411A7e866A631",
-    "spread": 0.01,
-    "maxApprovalExchangeRateChange": 0.5,
-    "vetoPeriodSeconds": 10800
-  },
-  "oracles": {
-    "reportExpiry": 300
-  },
-  "reserve": {
-    "initialBalance": 100000000,
-    "otherAddresses": ["0x91c987bf62D25945dB517BDAa840A6c661374402"]
-  },
-  "reserveSpenderMultiSig": {
-    "signatories": ["0x5409ed021d9299bf6814279a6a1411a7e866a631", "0x4404ac8bd8F9618D27Ad2f1485AA1B2cFD82482D"],
-    "numRequiredConfirmations": 2
-  },
-  "stableToken": {
-    "goldPrice": 1,
-    "initialBalances": {
-      "addresses": [
-        "0x5409ED021D9299bf6814279A6A1411A7e866A631",
-        "0x6Ecbe1DB9EF729CBe972C83Fb886247691Fb6beb",
-        "0xE36Ea790bc9d7AB70C55260C66D52b1eca985f84",
-        "0xE834EC434DABA538cd1b9Fe1582052B880BD7e63"
-      ],
-      "values": [
-        "50000000000000000000000",
-        "50000000000000000000000",
-        "50000000000000000000000",
-        "50000000000000000000000"
-      ]
-    },
-    "oracles": [
-      "0x5409ED021D9299bf6814279A6A1411A7e866A631",
-      "0xE36Ea790bc9d7AB70C55260C66D52b1eca985f84",
-      "0x06cEf8E666768cC40Cc78CF93d9611019dDcB628",
-      "0x7457d5E02197480Db681D3fdF256c7acA21bDc12"
-    ],
-    "frozen": false
-  },
-  "stableTokenEUR": {
-    "goldPrice": 1,
-    "initialBalances": {
-      "addresses": [
-        "0x5409ED021D9299bf6814279A6A1411A7e866A631",
-        "0x6Ecbe1DB9EF729CBe972C83Fb886247691Fb6beb",
-        "0xE36Ea790bc9d7AB70C55260C66D52b1eca985f84",
-        "0xE834EC434DABA538cd1b9Fe1582052B880BD7e63"
-      ],
-      "values": [
-        "50000000000000000000000",
-        "50000000000000000000000",
-        "50000000000000000000000",
-        "50000000000000000000000"
-      ]
-    },
-    "oracles": [
-      "0x5409ED021D9299bf6814279A6A1411A7e866A631",
-      "0xE36Ea790bc9d7AB70C55260C66D52b1eca985f84",
-      "0x06cEf8E666768cC40Cc78CF93d9611019dDcB628",
-      "0x7457d5E02197480Db681D3fdF256c7acA21bDc12"
-    ],
-    "frozen": false
-  },
-  "stableTokenBRL": {
-    "goldPrice": 1,
-    "initialBalances": {
-      "addresses": [
-        "0x5409ED021D9299bf6814279A6A1411A7e866A631",
-        "0x6Ecbe1DB9EF729CBe972C83Fb886247691Fb6beb",
-        "0xE36Ea790bc9d7AB70C55260C66D52b1eca985f84",
-        "0xE834EC434DABA538cd1b9Fe1582052B880BD7e63"
-      ],
-      "values": [
-        "50000000000000000000000",
-        "50000000000000000000000",
-        "50000000000000000000000",
-        "50000000000000000000000"
-      ]
-    },
-    "oracles": [
-      "0x5409ED021D9299bf6814279A6A1411A7e866A631",
-      "0xE36Ea790bc9d7AB70C55260C66D52b1eca985f84",
-      "0x06cEf8E666768cC40Cc78CF93d9611019dDcB628",
-      "0x7457d5E02197480Db681D3fdF256c7acA21bDc12"
-    ],
-    "frozen": false
-  },
-  "validators": {
-    "commissionUpdateDelay": 3
-  }
-}
