--- packages/dev-utils/tsconfig.json
@@ -1,10 +1,0 @@
-{
-  "extends": "@tsconfig/recommended/tsconfig.json",
-  "compilerOptions": {
-    "rootDir": "src",
-    "strict": false,
-    "outDir": "lib",
-    "resolveJsonModule": true
-  },
-  "include": ["src", "types/", "src/migration-override.json"]
-}
